Move HTTP requests to FileServlet for path traversal
