+++ src/main/webapp/xss.html
@@ -60,7 +60,6 @@
       <p class="test-class">
         At vero eos et accusamus et iusto odio dignissimos ducimus qui blanditiis
         praesentium voluptatum deleniti atque corrupti quos.
-        XSS
       </p>
       <h2>Second header h2</h2>
       <p class="test-class">
